Add check for dual architecture Python

+++ wheel_makers/make_wheels.sh
@@ -1,7 +1,14 @@
 # Make wheels and copy into main package
 # This is to build the wheels we use for testing
-# Need Cython and wheel installed to run this script
+# Need Cython and wheel installed to run this script.
+# Python needs to be dual architecture
 # Run on earliest supported version of macOS (currently 10.9)
+
+archs=$(lipo -archs `which python`)
+if [ "$archs" -ne "x86_64 arm64" ]; then
+    echo "Need dual architecture x86_64 arm64 Python"
+    exit 1
+fi
 
 rm */dist/fakepkg*.whl
 rm */libs/*.dylib
